// frontend/src/api/model/concourseItemCommentRead.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * Qualis API
 * OpenAPI spec version: 0.1.0
 */
import type { ConcourseItemCommentReadUserId } from './concourseItemCommentReadUserId';

/**
 * Schema for reading a comment on a concourse item.
 */
export interface ConcourseItemCommentRead {
    id: number;
    item_id: number;
    user_id?: ConcourseItemCommentReadUserId;
    body: string;
    created_at: string;
    updated_at: string;
}